Oops, the angle() method needs to return float rather than int32_t.
git-svn-id: http://svn/svn/hcr/trunk@1012 2e3f476d-dfe7-4101-bf40-6ec35b1080dd

+++ src/MotionControl/ElmoServoDrive.h
@@ -82,7 +82,7 @@ public:
     StatusReg driveStatusRegister() const { return _driveStatusRegister; }
 
     /// Get latest sampled drive angle, in degrees
-    int32_t angle() const { return(_angleCounts / countsPerDegree()); }
+    float angle() const { return(_angleCounts / countsPerDegree()); }
 
     /// Get latest sampled drive temperature, C
     int driveTemperature() const { return _driveTemperature; }
